added more delete functions and with-macros. Added a finalizers.lisp file and began filling with types/finalizers, updated library to include new type values in finalizers.lisp.  added nice example/documentation for descrip-matcher-match. added new length/size param, a keyword, to vector-float/int/double/char/uchar, call by (vector-float :size) for example

+++ src/size.cpp
@@ -5,7 +5,7 @@ extern "C" {
     Size* cv_create_Size() {
         return new Size;
     }
-    Size* cv_create_Size2(int width, int height) {
+    Size* cv_create_Size2(double width, double height) {
         return new Size(width, height);
     }
     Size* cv_Size_assignTo(Size* self, Size* other) {
